100% in the JaCoCo Tests lets gooo, fixedmost issues.
