+++ auth-service/src/test/java/com/petclinic/auth/AuthServiceUserControllerTests.java
@@ -17,12 +17,13 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 
+import javax.validation.ConstraintViolationException;
 import java.util.Random;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.instanceOf;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @AutoConfigureMockMvc
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -64,5 +65,38 @@ public class AuthServiceUserControllerTests {
         assertNotNull(user);
         assertThat(user.getId(), instanceOf(Long.TYPE));
         assertTrue(userRepo.findById(user.getId()).isPresent());
+    }
+    @Test
+    @DisplayName("Check the required fields with empty data")
+    void check_empty_require_fields() throws Exception{
+
+        UserIDLessDTO userIDLessDTO = new UserIDLessDTO();
+
+        assertThrows(ConstraintViolationException.class, () -> userController.createUser(userIDLessDTO));
+    }
+
+    @Test
+    @DisplayName("Check the username field in order to refused if it is empty")
+    void check_empty_username() throws Exception{
+
+        UserIDLessDTO userIDLessDTO = new UserIDLessDTO(null, PASS,EMAIL);
+
+        assertThrows(ConstraintViolationException.class, () -> userController.createUser(userIDLessDTO));
+    }
+    @Test
+    @DisplayName("Check the password field in order to refused if it is empty")
+    void check_empty_password() throws Exception{
+
+        UserIDLessDTO userIDLessDTO = new UserIDLessDTO( USER, null,EMAIL);
+
+        assertThrows(ConstraintViolationException.class, () -> userController.createUser(userIDLessDTO));
+    }
+    @Test
+    @DisplayName("Check the email field in order to refused if it is empty")
+    void check_empty_email() throws Exception{
+
+        UserIDLessDTO userIDLessDTO = new UserIDLessDTO(USER, PASS,null);
+
+        assertThrows(ConstraintViolationException.class, () -> userController.createUser(userIDLessDTO));
     }
 }
